pyproject.toml
[project]
name = "mynotes"
version = "1.0.0"
description = "A modern desktop notes application built with pywebview"
readme = "README.md"
requires-python = ">=3.7"
authors = [
    {name = "MyNotes Developer"}
]

dependencies = [
    "pywebview>=6.0",
    "jinja2>=3.1.2",
]

[project.optional-dependencies]
dev = [
    "black",
    "flake8",
    "pytest",
]

[tool.uv]
dev-dependencies = [
    "black",
    "flake8", 
    "pytest",
]

[tool.black]
line-length = 88
target-version = ['py37']
